+++ src/maxgradient/_simple_gradient.py
@@ -6,6 +6,7 @@ from operator import itemgetter
 from typing import Any, Dict, Generator, Iterable, List, Literal, Optional, Tuple
 
 import rich.style
+from cheap_repr import normal_repr, register_repr
 from maxgradient.color import Color, PyColorType
 from maxgradient.theme import GradientTheme
 from rich._pick import pick_bool
@@ -18,7 +19,6 @@ from rich.style import Style, StyleType
 from rich.text import Span, Text
 from rich.traceback import install as tr_install
 from snoop import spy
-from cheap_repr import register_repr, normal_repr
 
 GradientMethod = Literal["default", "list", "mono", "rainbow"]
 DEFAULT_JUSTIFY: JustifyMethod = "default"
@@ -53,10 +53,9 @@ class SimpleGradient(Text):
         "_style",
         "_spans",
         "end",
-        "cc",
+        "verbose"
     )
 
-    @spy
     def __init__(
         self,
         text: str | Text = "",
@@ -68,8 +67,10 @@ class SimpleGradient(Text):
         no_wrap: bool = False,
         style: StyleType = Style.null(),
         end: str = "",
-        spans: Optional[List[Span]] = None
+        spans: Optional[List[Span]] = None,
+        verbose: bool = False
     ) -> None:
+        self.verbose = verbose
         self.text = text
         _style = Style.parse(style) if isinstance(style, str) else style
 
@@ -138,9 +139,8 @@ class SimpleGradient(Text):
             self._text = sanitized_text
             self._spans = value.spans
         elif isinstance(value, str):
-            # if value == "":
-            #     te
-            #     raise ValueError("Text cannot be empty.")
+            if value == "":
+                raise ValueError("Text cannot be empty.")
             sanitized_text = strip_control_codes(value)
             self._length = len(sanitized_text)
             self._text = sanitized_text
@@ -169,7 +169,6 @@ class SimpleGradient(Text):
         else:
             self._style = Style.parse(style)
 
-    # @spy
     def generate_spans(self) -> Generator[Span, None, None]:
         """
         Generate the gradient's spans.
@@ -180,7 +179,7 @@ class SimpleGradient(Text):
         Returns:
             List[Span]: The gradient's spans
         """
-        if VERBOSE:
+        if self.verbose:
             console.log("Entered generate_gradient")
         triplet1 = self.color1.triplet
         r1, g1, b1 = triplet1
@@ -293,12 +292,13 @@ class SimpleGradient(Text):
     def as_text(self, style: StyleType, end: str = "") -> Text:
         return Text(self.plain, spans=self._spans, style=style, end=end)
 
+
 register_repr(SimpleGradient)(normal_repr)
 
 if __name__ == "__main__":  # pragma: no cover
     console = Console()
     console.line(2)
     sample_text = "SimpleGradient is a class that prints a `string` \nor `rich.text.Text` object as a gradient from \ncolor1 to color2 with an optional style."
-    gradient = SimpleGradient(sample_text, color1="green", color2="cyan", style="bold") # type: ignore
+    gradient = SimpleGradient(sample_text, color1="green", color2="cyan", style="bold")  # type: ignore
     gradient.highlight_regex(r"(`.+`)", "#af00ff")
     console.print(gradient, justify="center")