+++ tests/test_color.py
@@ -1,6 +1,12 @@
 import unittest
-
+from rich.color import blend_rgb
+from rich.color_triplet import ColorTriplet
 from maxgradient.color import Color
+from rich.console import Console
+from rich.traceback import install as tr_install
+
+console = Console()
+tr_install(console=console)
 
 
 class TestColor(unittest.TestCase):
@@ -18,55 +24,9 @@ class TestColor(unittest.TestCase):
         self.assertEqual(color._rgba.green, 0.0)
         self.assertEqual(color._rgba.blue, 0.0)
 
-    def test_color_creation_with_color_object(self):
-        original_color = Color("#ff0000")
-        color = Color(original_color)
-        self.assertEqual(color._original, "#ff0000")
-        self.assertEqual(color._rgba.red, 1.0)
-        self.assertEqual(color._rgba.green, 0.0)
-        self.assertEqual(color._rgba.blue, 0.0)
-        self.assertEqual(color._rgba, original_color._rgba)
-
-    def test_color_creation_with_rich_color_object(self):
-        original_color = Color("#ff0000").as_rich()
-        color = Color(original_color)
-        self.assertEqual(color._original, "#ff0000")
-        self.assertEqual(color._rgba.red, 1.0)
-        self.assertEqual(color._rgba.green, 0.0)
-        self.assertEqual(color._rgba.blue, 0.0)
-        self.assertEqual(color._rgba, original_color.triplet)
-
     def test_color_creation_with_invalid_value(self):
         with self.assertRaises(ValueError):
             Color(123)
-
-    def test_rich_representation(self):
-        color = Color("#ff0000")
-        rich_representation = color.__rich__()
-        expected_representation = (
-            "Color("
-            "[bold #ffffff]Color[bold #ffffff]"
-            "("
-            "[bold #ff0000]#ff0000[bold #ff0000]"
-            ")[bold #ffffff]"
-            ")"
-        )
-        self.assertEqual(rich_representation, expected_representation)
-
-    def test_as_rich(self):
-        color = Color("#ff0000")
-        rich_color = color.as_rich()
-        self.assertEqual(rich_color.triplet, color._rgba)
-
-    def test_as_style(self):
-        color = Color("#ff0000")
-        style = color.as_style()
-        self.assertEqual(style.color.triplet, color._rgba)
-
-    def test_as_bg_style(self):
-        color = Color("#ff0000")
-        bg_style = color.as_bg_style()
-        self.assertEqual(bg_style.bgcolor.triplet, color._rgba)
 
     def test_get_contrast(self):
         color = Color("#ff0000")
@@ -79,12 +39,25 @@ class TestColor(unittest.TestCase):
         self.assertEqual(alpha_style, color._rgba.as_triplet())
 
     def test_get_alpha_style_with_bg_color(self):
-        color = Color("#ff0000")
-        bg_color = Color("#0000ff")
+        color = Color("rgb(255,0,0,0.5)")
+        bg_color = Color("#000000")
         alpha_style = color.get_alpha_style(bg_color)
-        expected_alpha_style = color._rgba.blend(bg_color._rgba)
-        self.assertEqual(alpha_style, expected_alpha_style)
-
-
-if __name__ == "__main__":
-    unittest.main()
+        expected_alpha_style = blend_rgb(
+            ColorTriplet(255,0,0),
+            ColorTriplet(0,0,0),
+            0.5
+        )
+def test_rich_representation(self):
+    color = Color("#ff0000")
+    rich_representation = color.__rich__()
+    expected_representation = (
+        "Color("
+        "[bold #ffffff]Color[bold #ffffff]"
+        "("
+        "[bold #ff0000]#ff0000[bold #ff0000]"
+        ")[bold #ffffff]"
+        ")"
+    )
+    print("rich_representation:", rich_representation)
+    print("expected_representation:", expected_representation)
+    self.assertEqual(rich_representation, expected_representation)